poor but working decoder

+++ discrete_val.h
@@ -5,12 +5,12 @@
 
 class discrete_value {
 private:
-std::map<float, float> distrib;
+std::map<double, double> distrib;
 
 public:
     discrete_value() : distrib({}) {}
 
-    discrete_value(std::map<float, float> distr) : distrib(distr) {}
+    discrete_value(std::map<double, double> distr) : distrib(distr) {}
 
     void print() {
         for (auto it = distrib.begin(); it != distrib.end(); ++it) {
@@ -22,11 +22,11 @@ public:
     }
 
     discrete_value sum_distrib(discrete_value other) {
-        std::map<float, float> help;
+        std::map<double, double> help;
         for (auto it1 = distrib.begin(); it1 != distrib.end(); ++it1) {
             for (auto it2 = other.distrib.begin(); it2 != other.distrib.end(); ++it2) {
-                float key = it1->first + it2->first;
-                float val = it1->second * it2->second;
+                double key = it1->first + it2->first;
+                double val = it1->second * it2->second;
                 if (help.find(key) != help.end()) {
                     help[key] += val;
                 } else {
@@ -38,11 +38,11 @@ public:
     }
 
     discrete_value max_distrib(discrete_value other) {
-        std::map<float, float> help;
+        std::map<double, double> help;
         for (auto it1 = distrib.begin(); it1 != distrib.end(); ++it1) {
             for (auto it2 = other.distrib.begin(); it2 != other.distrib.end(); ++it2) {
-                float key = std::max(it1->first, it2->first);
-                float val = it1->second * it2->second;
+                double key = std::max(it1->first, it2->first);
+                double val = it1->second * it2->second;
                 if (help.find(key) != help.end()) {
                     help[key] += val;
                 } else {
@@ -54,11 +54,11 @@ public:
     }
 
     discrete_value min_distrib(discrete_value other) {
-        std::map<float, float> help;
+        std::map<double, double> help;
         for (auto it1 = distrib.begin(); it1 != distrib.end(); ++it1) {
             for (auto it2 = other.distrib.begin(); it2 != other.distrib.end(); ++it2) {
-                float key = std::min(it1->first, it2->first);
-                float val = it1->second * it2->second;
+                double key = std::min(it1->first, it2->first);
+                double val = it1->second * it2->second;
                 if (help.find(key) != help.end()) {
                     help[key] += val;
                 } else {
@@ -70,11 +70,11 @@ public:
     }
 
     discrete_value signed_min_abs_distrib(discrete_value other) {
-        std::map<float, float> help;
+        std::map<double, double> help;
         for (auto it1 = distrib.begin(); it1 != distrib.end(); ++it1) {
             for (auto it2 = other.distrib.begin(); it2 != other.distrib.end(); ++it2) {
-                float key = (it1->first * it2->first / (abs(it1->first) * abs(it2->first))) * std::min(abs(it1->first), abs(it2->first));
-                float val = it1->second * it2->second;
+                double key = (it1->first * it2->first / (abs(it1->first) * abs(it2->first))) * std::min(abs(it1->first), abs(it2->first));
+                double val = it1->second * it2->second;
                 if (help.find(key) != help.end()) {
                     help[key] += val;
                 } else {
